feat: Add my task 1 after wrong merge & clean my project from initial code getting from main
