--- packages/backend/.env.example
@@ -1,2 +1,0 @@
-DATABASE_URL=database_url
-PORT=port
