+++ packages/backend/prisma/schema.prisma
@@ -1,4 +1,4 @@
-// This is your Prisma schema file,
+// This is your Prisma schemas file,
 // learn more about it in the docs: https://pris.ly/d/prisma-schema
 
 // Looking for ways to speed up your queries, or scale easily with your serverless or edge functions?
@@ -22,12 +22,13 @@ model Todo {
   createdAt   DateTime @default(now())
   updatedAt   DateTime @default(now()) @updatedAt
 }
+
 model TodoItem {
-  id            Int       @id @default(autoincrement())
-  title         String
-  description   String
-  isCompleted   Boolean   @default(false)
-  isPrivate     Boolean   @default(true)
-  createdAt     DateTime  @default(now())
-  updatedAt     DateTime  @updatedAt
+  id          Int      @id @default(autoincrement())
+  title       String
+  description String
+  isCompleted Boolean  @default(false)
+  isPrivate   Boolean  @default(true)
+  createdAt   DateTime @default(now())
+  updatedAt   DateTime @updatedAt
 }
