--- packages/backend/src/services/prisma/prisma.service.ts
@@ -1,2 +1,0 @@
-import { PrismaClient } from '@prisma/client';
-export const prisma = new PrismaClient();
