--- packages/backend/src/types/prisma.types.ts
@@ -1,7 +1,0 @@
-import { PrismaPromise } from '@prisma/client';
-
-export interface PrismaModelWithId<T> {
-	findUniqueOrThrow: (args: {
-		where: { id: number };
-	}) => PrismaPromise<T | null>;
-}
